+++ features/step_definitions/cli_steps.ts
@@ -76,22 +76,6 @@ When(
   }
 )
 
-When(
-  'I run cucumber-js \\(installed locally\\)',
-  { timeout: 10000 },
-  async function (this: World) {
-    return await this.run(this.localExecutablePath, [])
-  }
-)
-
-When(
-  'I run cucumber-js \\(installed globally\\)',
-  { timeout: 10000 },
-  async function (this: World) {
-    return await this.run(this.globalExecutablePath, [])
-  }
-)
-
 Then('it passes', () => {}) // eslint-disable-line @typescript-eslint/no-empty-function
 
 Then('it fails', function (this: World) {