+++ src/api/run_cucumber.ts
@@ -55,8 +55,10 @@ export async function runCucumber(
 
   let formatterStreamError = false
   const cleanup = await initializeFormatters({
+    env,
     cwd,
     stdout,
+    stderr,
     logger,
     onStreamError: () => (formatterStreamError = true),
     eventBroadcaster,