+++ src/formatter/publish.ts
@@ -1,2 +1,7 @@
+export interface IPublishConfig {
+  url: string
+  token: string
+}
+
 export const DEFAULT_CUCUMBER_PUBLISH_URL =
   'https://messages.cucumber.io/api/reports'